requirements.txt
Flask==3.0.0
flask[async]==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Caching==2.3.0
flask-marshmallow==1.2.1
psycopg2==2.9.10
marshmallow-sqlalchemy==0.28.0
python-json-logger==2.0.7
gunicorn==23.0.0
requests==2.32.3
beautifulsoup4==4.12.3
aiohttp==3.11.5
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2